Configured gateway for role based access

+++ Microservices/user-create-service/src/main/java/com/sapient/usercreateservice/entities/Users.java
@@ -13,6 +13,6 @@ public interface Users {
     String userId();
     String password();
     Optional<String> fullName();
-    Optional<Double> currBal();
+    Optional<Float> currBal();
     Optional<String> role();
 }
